quant-v4lite/requirements.txt
# 交易所
ccxt>=4.0.0
websockets>=12.0
python-binance>=1.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 数据库
asyncpg>=0.28.0
redis>=5.0.0
sqlalchemy>=2.0.0

# 配置
pyyaml>=6.0
python-dotenv>=1.0.0

# 通知
python-telegram-bot>=20.0

# 回测
matplotlib>=3.7.0
tabulate>=0.9.0

# 工具
aiohttp>=3.9.0
loguru>=0.7.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0
